requirements.txt
click>=8.1.8
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai==0.130.0 
crewai-tools==0.47.1
fastapi==0.110.3
google-ai-generativelanguage==0.6.4
google-api-core>=2.11.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
Jinja2==3.1.4
jsonschema==4.22.0
langchain-core==0.1.52
langchain-community>=0.0.38
langsmith==0.1.67
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai>=1.68.2
opentelemetry-api>=1.30.0
opentelemetry-exporter-otlp-proto-common>=1.30.0
opentelemetry-exporter-otlp-proto-grpc>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-proto
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
pandas==2.2.2
pillow==10.3.0
protobuf>=4.25.8
pydantic>=2.4.2
pydantic_core==2.8.0
python-dotenv>=1.0.0
pypdf>=3.0.0
uvicorn>=0.29.0
python-multipart>=0.0.9
celery>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0